% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPhylopicTreePBDB.R
\name{cachePhyloPicPNG}
\alias{cachePhyloPicPNG}
\title{Cache phylopic images for a tree}
\usage{
cachePhyloPicPNG(tree, taxaDataPBDB = tree$taxaDataPBDB, cacheDir)
}
\arguments{
\item{tree}{A phylogeny of class \code{phylo} which will be
plotted, with the terminal tip taxa replaced by silhouettes.
The tree will be plotted with edge lengths.}

\item{taxaDataPBDB}{A \code{data.frame} of taxonomic data from
the Paleobiology Database containing an \code{$image_no} variable,
as returned when \code{show = "img"} is used. See \emph{Details}.}

\item{cacheDir}{Where to save the output PNGs}
}
\description{
This function takes a tree and saves phylopic thumbnails locally,
in order to dramatically speed up later rendering.
}
