% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exhaustationFunctions.R
\name{exhaustationFunctions}
\alias{exhaustationFunctions}
\alias{accioExhaustionCurve}
\alias{accioBestAcquisitionModel}
\alias{charExhaustPlot}
\title{Exhaustation of Characters}
\usage{
accioExhaustionCurve(phyloTree, charData, charTypes, outgroup = NULL,
  firstAppearances = NULL, missingCharValue = "?",
  inapplicableValue = "-")

accioBestAcquisitionModel(changes, models = c("exponential", "gamma",
  "lognormal", "zipf"))

charExhaustPlot(exhaustion_info, plotType, xlab = "Total Characters",
  ylab = NULL, main = NULL, xsize = 3)
}
\arguments{
\item{phyloTree}{A phylogenetic tree of class \code{phylo} as used by package \code{ape}.}
}
\value{

}
\description{
Exhaustation of Characters
}
\examples{

# get data
data(SongZhangDicrano)

phyloTree<-cladogramDicranoX13

# stratigraphic age data
strat_data<-cbind(c(0, 4, 7, 7, 6, 6, 6, 2, 2, 1, 1, 6, 7), c(0, 5, 7, 7, 7, 7, 7, 2, 2, 1, 1, 7, 7))
# get just first appearances from strat data
FAs <- strat_data[,1]
 
# modify char data
charMat<-data.matrix(charMatDicrano)
charMat[is.na(charMatDicrano)]<-0
charMat<-(charMat-1)
colnames(charMat)<-NULL
charMat[is.na(charMatDicrano)]<-"?"

# get character types
nchar <- dim(charMat)[2]
types <- rep(0,nchar)	# set character types to unordered (1)


#the 'outgroup' is Exigraptus, first taxon listed in the matrix
exhaustionResults <- accioExhaustionCurve(phyloTree,charMat,types,FAs,outgroup=1)




# plot of exhausation of total accumulation of character states

charExhaustPlot(exhaustion_info=exhaustionResults,
	plotType="totalAcc")

	
# plot of exhausation of character alterations

charExhaustPlot(exhaustion_info=exhaustionResults,
	plotType="charAlt")
	
}
\references{
Wagner, P. J. 2000. Exhaustion of morphologic character states among fossil taxa. Evolution 54(2):365-386.
}
\seealso{

}
\author{
David W. Bapst
}
