% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTaxaDataPBDB.R
\name{getTaxaDataPBDB}
\alias{getTaxaDataPBDB}
\alias{getCladeTaxaPBDB}
\alias{getSpecificTaxonTreePBDB}
\alias{getSpecificTaxonTreePBDB}
\title{Obtaining Data for Sets of Taxa From Paleobiology Database API}
\usage{
getCladeTaxaPBDB(taxon, show = c("class", "img", "app", "parent"),
  status = "accepted")

getSpecificTaxonTreePBDB(taxa, PDFfile = "myPBDBphylogeny.pdf",
  show = c("class", "img", "app", "parent"), status = "accepted",
  addressOnly = FALSE)
}
\arguments{
\item{taxon}{A single name of a of a higher taxon which you wish to catch
all taxonomic 'children' (included members - i.e. subtaxa) of, from 
within the Paleobiology Database.}
}
\description{
The Paleobiology Database API () is very easy to use, and generally any data one wishes to collect
}
\examples{

#an example here
}
\references{
Peters, S. E., and M. McClennen. 2015. The Paleobiology Database
application programming interface. \emph{Paleobiology} 42(1):1-7.
}
\seealso{
See \code{\link{getTaxaDataPBDB}}, \code{\link{makePBDBtaxonTree}},
and \code{\link{plotPhylopicTreePBDB}}.
}
\author{
David W. Bapst
}
