% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/makePBDBtaxontree.R
\name{makePBDBtaxontree}
\alias{makePBDBtaxontree}
\title{Creating a Taxon-Tree from Taxonomic Data Downloaded from the Paleobiology Database}
\usage{
makePBDBtaxontree(data, rank)
}
\arguments{
\item{data}{A table of taxonomic data collected from the Paleobiology Database, using the taxa list option
with show=phylo.}

\item{rank}{The selected taxon rank; must be one of 'species', 'genus', 'family', 'order', 'class' or 'phylum'.}
}
\value{
A phylogeny of class 'phylo', where each tip is a taxon of the given 'rank'. Edges are scaled so that
the distance from one taxon rank to another 1, then merged to remove singleton nodes. As not all
taxa have parents at the immediate taxon level above, this leads to some odd cases. For example,
two genera emanating from a node representing a class but with a very short (length=1) branch
and a long branch (length=3) means one genus is simply placed in the class, with no family or order listed
wile the one on the long branch is within an order and family that is otherwise monogeneric.
}
\description{
This function creates phylogeny-like object of type phylo from the taxonomic hierarchy
recorded in a taxonomy download from the PBDB for a given group.
}
\details{
This function should not be taken too seriously. Many groups in the Paleobiology Database have
out-of-date or very incomplete taxonomic information. This function is meant to help visualize
what information is present, and by use of time-scaling functions, allow us to visualize the intersection
of temporal and phylogenetic, mainly to look for incongruence due to either incorrect taxonomic placements,
erroneous occurrence data or both.

Note however that, contrary to common opinion among some paleontologists, taxon-trees may be just as useful for
macroevolutionary studies as reconstructed phylogenies (Soul and Friedman, in press.)
}
\examples{
\dontrun{

getTaxaPBDB<-function(taxon){
	#let's get some taxonomic data
	taxaData<-read.csv(paste0("http://paleobiodb.org/",
		"data1.1/taxa/list.txt?base_name=",taxon,
		"&rel=all_children&show=phylo,img,nav&status=senior"))
	return(taxaData)
	}

#graptolites
data<-getTaxaPBDB("Graptolithina")
graptTree<-makePBDBtaxontree(data,"genus")
plot(graptTree)

#conodonts
data<-getTaxaPBDB("Conodonta")
conoTree<-makePBDBtaxontree(data,"genus")
plot(conoTree)

#asaphid trilobites
data<-getTaxaPBDB("Asaphida")
asaTree<-makePBDBtaxontree(data,"genus")
plot(asaTree)

}

###################################

#let's try time-scaling the graptolite tree

#get some example occurrence and taxonomic data
data(graptPBDB)

#get taxontree
graptTree<-makePBDBtaxontree(graptTaxaPBDB,"genus")
plot(graptTree)

#get time data from occurrences
graptOccGenus<-taxonSortPBDBocc(graptOccPBDB,rank="genus",onlyFormal=FALSE)
graptTimeGenus<-occData2timeList(occList=graptOccGenus)

#let's time-scale this tree with paleotree
timeTree<-bin_timePaleoPhy(graptTree,timeList=graptTimeGenus,
	nonstoch.bin=TRUE,type="mbl",vartime=3)

#drops a lot of taxa; some of this is due to mispellings, etc

\dontrun{

#make pretty plot with library strap
library(strap)
geoscalePhylo(timeTree, ages=timeTree$ranges.used)

}
}
\author{
David W. Bapst
}
\references{
Soul, L. C., and M. Friedman. In Press. Taxonomy and Phylogeny Can Yield
Comparable Results in Comparative Palaeontological Analyses.\emph{Systematic Biology}
(http://sysbio.oxfordjournals.org/content/early/2015/03/23/sysbio.syv015.abstract)
}
\seealso{
Other functions for manipulating PBDB data can be found at \code{\link{taxonSortPBDBocc}},
\code{\link{occData2timeList}}, and the example data at \code{\link{graptPBDB}}.
}

