% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/occData2timeList.R
\name{occData2timeList}
\alias{occData2timeList}
\title{Converting Occurrences Data to a timeList Data Object
This function converts occurrence data, given as a list where each element
is a different taxon's occurrence table (containing minimum and maximum ages
for each occurrence), to the 'timeList' format, consisting of a list consisting
of a matrix of lower and upper age bounds for intervals and a second matrix
recording the interval in which taxa first and last occur in the given dataset.}
\usage{
occData2timeList(occList, intervalType = "dateRange")
}
\arguments{
\item{occList}{A list where every element is a table of occurrence data for a different taxon,
such as that returned by \code{\link{taxonSortPBDBocc}}. The occurrence data can be either a
two-column matrix composed of the lower and upper age bounds on each taxon occurrence, or has
two named variables which match either the}

\item{intervalType}{Must be either "dateRange" (the default) or "occRange". If "dateRange",
the bounds on the first appearances are the span between the oldest upper and lower bounds
of the occurrences, and the bounds on the last appearances are the span between the youngest
upper and lower bounds across all occurrences. This is guaranteed to provide the smallest
bounds on the first and last appearances, and was suggested by Jon Marcot. The "occRange" option
instead returns the smallest bounds among (a) the oldest occurrences for the first appearance
(i.e. all occurrences with their lowest bound at the oldest lower age bound), and (b) the
youngest occurrences for the last appearance (i.e. all occurrences with their uppermost bound
at the youngest upper age bound).}
}
\value{
Returns a standard timeList data object, as used by many other paleotree functions, like \code{\link{bin_timePaleoPhy}}, \code{\link{bin_cal3TimePaleoPhy}} and \code{\link{taxicDivDisc}}
}
\description{
Converting Occurrences Data to a timeList Data Object
This function converts occurrence data, given as a list where each element
is a different taxon's occurrence table (containing minimum and maximum ages
for each occurrence), to the 'timeList' format, consisting of a list consisting
of a matrix of lower and upper age bounds for intervals and a second matrix
recording the interval in which taxa first and last occur in the given dataset.
}
\details{
This function should translate taxon-sorted occurrence data (i.e. datasets output by \code{\link{taxonSortPBDBocc}}
}
\examples{
data(graptPBDB)

graptOccSpecies<-taxonSortPBDBocc(graptOccPBDB,rank="species",onlyFormal=FALSE)
graptTimeSpecies<-occData2timeList(occList=graptOccSpecies)

head(graptTimeSpecies[[1]])
head(graptTimeSpecies[[2]])

graptOccGenus<-taxonSortPBDBocc(graptOccPBDB,rank="genus",onlyFormal=FALSE)
graptTimeGenus<-occData2timeList(occList=graptOccGenus)

layout(1:2)
taxicDivDisc(graptTimeSpecies)
taxicDivDisc(graptTimeGenus)

# the default interval calculation is "dateRange"
# let's compare to the other option, "occRange"
	#for species

graptOccRange<-occData2timeList(occList=graptOccSpecies, intervalType="occRange")

#we would expect no change in the diversity curve
	#because there are only changes in th
		#earliest bound for the FAD
		#latest bound for the LAD
#so if we are depicting ranges within maximal bounds
	#dateRanges has no effect
layout(1:2)
taxicDivDisc(graptTimeSpecies)
taxicDivDisc(graptOccRange)
#yep, identical

#so how much uncertainty was gained by using dateRange?

# write a simple function for getting uncertainty in first and last
		# appearance dates from a timeList object
sumAgeUncert<-function(timeList){
	fourDate<-timeList2fourDate(timeList)
	perOcc<-(fourDate[,1]-fourDate[,2])+(fourDate[,3]-fourDate[,4])
	sum(perOcc)
	}

#total amount of uncertainty in occRange dataset
sumAgeUncert(graptOccRange)
#total amount of uncertainty in dateRange dataset
sumAgeUncert(graptTimeSpecies)
#the difference
sumAgeUncert(graptOccRange)-sumAgeUncert(graptTimeSpecies)
#as a proportion
1-(sumAgeUncert(graptTimeSpecies)/sumAgeUncert(graptOccRange))

#a different way of doing it
dateChange<-timeList2fourDate(graptTimeSpecies)-timeList2fourDate(graptOccRange)
apply(dateChange,2,sum)
#total amount of uncertainty removed by dateRange algorithm
sum(abs(dateChange))

layout(1)
}
\author{
David W. Bapst, with the 'dateRange' algorithm suggested by Jon Marcot.
}
\seealso{
\code{\link{taxonSortPBDBocc}} and the example graptolite dataset at \code{\link{graptPBDB}}
}

