% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parentChild2TaxonTree.R
\name{parentChild2TaxonTree}
\alias{parentChild2TaxonTree}
\title{Create a Taxonomy-Based Phylogeny ('Taxon Tree') from a Table of Parent-Child Taxon Relationships}
\usage{
parentChild2TaxonTree(parentChild, tipSet = "nonParents", cleanTree = TRUE)
}
\arguments{
\item{parentChild}{A two-column matrix of type \code{character} where
each element is a taxon name. Each row represents a parent-child relationship
with first the parent (column 1) taxon name and then the child (column 2).}

\item{tipSet}{This argument controls which taxa are selected as tip taxa for the
output tree. The default \code{tipSet="nonParents"} selects all child taxa which
are not listed as parents in \code{parentChild}. Alternatively, \code{tipSet="all"}
will add a tip to every internal node with the parent-taxon name encapsulated in
parentheses.}

\item{cleanTree}{By default, the tree is run through a series of post-processing, including having singles collapsed,
nodes reordered and being written out as a Newick string and read back in, to ensure functionality with ape functions
and ape-derived functions. If FALSE, none of this post-processing is done and users should beware, as such trees can
lead to hard-crashes of R.}
}
\value{
A phylogeny of class 'phylo', with tip taxa as controlled by argument \code{tipSet}.
The output tree is returned with no edge lengths.

The names of higher taxa than the tips should be appended as the element $node.label for the internal nodes.
}
\description{
This function takes a two-column matrix of taxon names,
indicating a set of binary parent-taxon:child-taxon
paired relationships with a common root, and returns
a 'taxonomy-tree' phylogeny object of class 'phylo'.
}
\details{
All taxa listed must be traceble via their parent-child relationships to a single,
common ancestor which will act as the root node for output phylogeny.
}
\examples{
#let's create a small, really cheesy example
pokexample<-rbind(cbind("Squirtadae",c("Squirtle","Blastoise","Wartortle")),
	c("Shelloidea","Lapras"),c("Shelloidea","Squirtadae"),
	c("Pokezooa","Shelloidea"),c("Pokezooa","Parasect"),
	c("Rodentapokemorpha","Linoone"),c("Rodentapokemorpha","Sandshrew"),
	c("Rodentapokemorpha","Pikachu"),c("Hirsutamona","Ursaring"),
	c("Hirsutamona","Rodentapokemorpha"),c("Pokezooa","Hirsutamona"))

#Default: tipSet='nonParents'
pokeTree<-parentChild2TaxonTree(pokexample, tipSet="nonParents")
plot(pokeTree);nodelabels(pokeTree$node.label)

#Get ALL taxa as tips with tipSet='all'
pokeTree<-parentChild2TaxonTree(pokexample, tipSet="all")
plot(pokeTree);nodelabels(pokeTree$node.label)


\dontrun{

# let's try a dataset where not all the taxon relationships lead to a common root

pokexample_bad<-rbind(cbind("Squirtadae",c("Squirtle","Blastoise","Wartortle")),
	c("Shelloidea","Lapras"),c("Shelloidea","Squirtadae"),
	c("Pokezooa","Shelloidea"),c("Pokezooa","Parasect"),
	c("Rodentapokemorpha","Linoone"),c("Rodentapokemorpha","Sandshrew"),
	c("Rodentapokemorpha","Pikachu"),c("Hirsutamona","Ursaring"),
	c("Hirsutamona","Rodentapokemorpha"),c("Pokezooa","Hirsutamona"),
	c("Umbrarcheota","Gengar"))

#this should return an error, as Gengar doesn't share common root
pokeTree<-parentChild2TaxonTree(pokexample_bad)

}
}
\author{
David W. Bapst
}
\seealso{
\code{\link{makePBDBtaxontree}}, \code{\link{taxonTable2TaxonTree}}
}

