% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPhylopicTreePBDB.R
\name{plotPhylopicTreePBDB}
\alias{plotPhylopicTreePBDB}
\title{Plot a Phylogeny with Organismal Silhouettes from PhyloPic, Called Via the Paleobiology Database API}
\usage{
plotPhylopicTreePBDB(tree, taxaDataPBDB = taxaTree$taxaData, size = 0.9,
  noiseThreshold = 0.1, extraMargin = 0.2, rescalePNG = TRUE,
  trimPNG = TRUE)
}
\arguments{
\item{tree}{}

\item{size}{= 0.9}

\item{noiseThreshold}{= 0.1}

\item{extraMargin}{= 0.2}

\item{rescalePNG}{= TRUE}

\item{trimPNG}{= TRUE}

\item{phylopicIDsPBDB}{NULL}
}
\value{
This function returns nothing - it just plots the input tree with
silhouettes from the PhyloPic database.
}
\description{
Plot a Phylogeny with Organismal Silhouettes from PhyloPic, Called Via the Paleobiology Database API
}
\details{

}
\examples{

\donttest{


# now plot with phylopic images
plotPhylopicTreePBDB(tree = tree, 
	taxaDataPBDB = taxaData)

}

}
\references{
Peters, S. E., and M. McClennen. 2015. The Paleobiology Database
application programming interface. \emph{Paleobiology} 42(1):1-7.
}
\seealso{
See \code{\link{getTaxaDataPBDB}}, \code{\link{makePBDBtaxonTree}},
and \code{\link{plotPhylopicTreePBDB}}.
}
\author{
David W. Bapst
}
