% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simFossilRecordMethods.R
\name{simFossilRecordMethods}
\alias{fossilRecord2fossilRanges}
\alias{fossilRecord2fossilTaxa}
\alias{simFossilRecordMethods}
\alias{timeSliceFossilRecord}
\title{Methods for Editing or Converting Output from simFossilRecord}
\usage{
timeSliceFossilRecord(fossilRecord, sliceTime, shiftRoot4TimeSlice = FALSE,
  modern.samp.prob = 1, tolerance = 10^-4)

fossilRecord2fossilTaxa(fossilRecord)

fossilRecord2fossilRanges(fossilRecord, merge.cryptic = TRUE,
  ranges.only = TRUE)
}
\arguments{
\item{fossilRecord}{A list object output by \code{simFossilRecord}, often composed
of multiple elements, each of which is data for 'one taxon', with the first
element being a distinctive six-element vector composed of numbers, corresponding
to the six numbers in a \code{simFossilTaxa} matrix.}

\item{sliceTime}{The date to slice the \code{simFossilRecord} output at, given
in time-units before the modern, on the same scale as the input \code{fossilRecord}.}

\item{shiftRoot4TimeSlice}{Should the dating of events be shifted, so that the
date given for \code{sliceTime} is now 0, or should the dates not be shifted,
so that they remain on the same scale as the input? This argument accepts a
logical TRUE or FALSE, but also accepts the string \code{"withExtantOnly"},
which will only 'shift' the time-scale if living taxa are present, as
determined by having ranges that overlap within \code{tolerance} of \code{sliceTime}.}

\item{modern.samp.prob}{Probability of sampling living taxa at the present
day (time=0), see below.}

\item{tolerance}{A small number which sets a range around the \code{sliceTime} within
which taxa will be considered extant.}

\item{merge.cryptic}{If TRUE, sampling events for cryptic species will be
merged into one taxon.}

\item{ranges.only}{If TRUE, gives taxon first and last occurrences only. If
FALSE, gives the time of all sampling events as a list.}
}
\value{
Depends on the function.
}
\description{
Methods for Editing or Converting Output from simFossilRecord
}
\details{
These functions exist to manipulate output from \code{simFossilRecord},
particularly so that
}
\examples{
set.seed(444)
record <- simFossilRecord(p=0.1, q=0.1, r=0.1, nruns=1,
	nTotalTaxa=c(20,30) ,nExtant=0, plot=TRUE)
}
\author{
David W. Bapst.
}
\seealso{
\code{\link{simFossilRecord}}
}

