% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateTaxonTreePBDB.R
\name{dateTaxonTreePBDB}
\alias{dateTaxonTreePBDB}
\title{Date a Taxon-Topology from the Paleobiology Database Using Appearance Data from the API}
\usage{
dateTaxonTreePBDB(taxaTree, taxaDataPBDB = taxaTree$taxaDataPBDB,
  minBranchLen = 0, plot = FALSE)
}
\arguments{
\item{taxaTree}{A tree with tip taxon names matching the taxon names
in \code{taxaDataPBDB}. Probably a taxon tree estimated
with \code{\link{makePBDBtaxonTree}}.}

\item{taxaDataPBDB}{A data table of taxonomic information obtained
using the Paleobiology Database's API for a set of taxa that
includes the tip taxa on \code{taxaTree}, generated with
parameter \code{show=app} so that appearance times are included.}

\item{minBranchLen}{Following dating using the appearance times taken directly
from the PBDB for each tip taxon and node, the tree may then be assessed with
the minimum branch length algorithm, as applied by \code{\link{minBranchLength}}.
If \code{minBranchLen = 0}, the default, this step is skipped. It may be necessary
to set \code{minBranchLen} higher than zero to differentiate nodes in cases
with poor stratigraphic congruency, so that derived taxa are the first taxa
observed in a group.}

\item{plot}{If \code{TRUE}, the resulting dated tree is plotted.}
}
\value{
Returns a dated phylogeny of class \code{phylo}, with an additional element
\code{$taxaDataPBDB} added containing the input \code{taxaDataPBDB}, as this might be
called by other functions.
}
\description{
The function \code{dateTaxonTreePBDB} takes a topology of  The required input is a topology with tip and internal node labels corresponding to taxa in the Paleobiology Database, and a table of data (containing those same tip and node taxa) obtained from the taxa-list functionality of the Paleobiology Database's API, with appearance times output.
}
\details{
The dating by
}
\examples{

#an example here
}
\references{
Peters, S. E., and M. McClennen. 2015. The Paleobiology Database
application programming interface. \emph{Paleobiology} 42(1):1-7.
}
\seealso{
See \code{\link{gettaxaDataPBDB}}, \code{\link{makePBDBtaxonTree}},
and \code{\link{plotPhylopicTreePBDB}}.
}
\author{
David W. Bapst
}
