% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTaxaDataPBDB.R
\name{getTaxaDataPBDB}
\alias{getTaxaDataPBDB}
\alias{getCladeTaxaPBDB}
\alias{getSpecificTaxaPBDB}
\alias{getSpecificTaxaPBDB}
\title{Obtaining Data for Sets of Taxa From Paleobiology Database API}
\usage{
getCladeTaxaPBDB(taxon, show = c("class", "parent", "app", "img", "entname"),
  status = "accepted", urlOnly = FALSE)

getSpecificTaxaPBDB(taxa, show = c("class", "parent", "app", "img",
  "entname"), status = "accepted", urlOnly = FALSE)
}
\arguments{
\item{taxon}{A single name of a of a higher taxon which you wish to catch
all taxonomic 'children' (included members - i.e. subtaxa) of, from 
within the Paleobiology Database.}

\item{show}{Which variables should be requested from the Paleobiology
Database? The default is to include classification (\code{"class"}),
parent-child taxon information (\code{"parent"}), information on each
taxon's first and last appearance (\code{"app"}), information on the
PhyloPic silhouette images assigned to that taxon (\code{"img"}), and
the names of those who entered and authorized the taxonomic data you
have downloaded (\code{"entname"}). Multiple variable blocks can be
given as a single character string, with desired variable selections
separated by a comma with no whitespace (ex. \code{"class,img,app"}) or 
as a vector of character strings (ex. \code{c("class", "img", "app")}),
which will then formatted for use in the API call. Other options that
you might want to include, such as information on ecospace or taphonomy,
can be included: please refer to the full list at
the \href{http://paleobiodb.org/data1.2/taxa/list_doc.htm}{documentation for the API}.}

\item{status}{What taxonomic status should the pull taxa have? 
The default is \code{status = "accepted"}, which means 
only those taxa that are both valid taxa and 
\emph{the accepted senior homonym}. Other typical statuses
to consider are \code{"valid"}, which is all valid taxa:
senior homonyms and valid subjective synonyms, and \code{"all"},
which will return all valid taxa and all otherwise repressed invalid taxa.
For additional statuses that you can request, please see the documentation at the 
\href{http://paleobiodb.org/data1.2/taxa/list_doc.htm}{documentation for the API}.}

\item{urlOnly}{If \code{FALSE} (the default), then the
function behaves as expected, the API is called and a
data table pulled from the Paleobiology Database is returned.
If \code{urlOnly = TRUE}, the URL of the API call is returned
instead as a character string.}

\item{taxa}{A character vector listing taxa of interest that the user
wishes to download information on from the Paleobiology Database.
Multiple taxa can be listed as a single character string, with desired taxa
separated by a comma with no whitespace (ex.
\code{"Homo,Pongo,Gorilla"}) or as a vector of character strings
(ex. \code{c("Homo", "Pongo", "Gorilla")}),
which will then formatted for use in the API call.}
}
\value{
These functions return a \code{data.frame} containing
variables pulled for the requested taxon selection.
This behavior can be modified by argument \code{urlOnly}.
}
\description{
The Paleobiology Database API () is very easy to use, and generally any data one wishes to collect
}
\details{
The
}
\examples{

#an example here
}
\references{
Peters, S. E., and M. McClennen. 2015. The Paleobiology Database
application programming interface. \emph{Paleobiology} 42(1):1-7.
}
\seealso{
See \code{\link{getTaxaDataPBDB}}, \code{\link{makePBDBtaxonTree}},
and \code{\link{plotPhylopicTreePBDB}}.
}
\author{
David W. Bapst
}
