% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtainDatedPosteriorTreesMrB.R
\name{obtainDatedPosteriorTreesMrB}
\alias{obtainDatedPosteriorTreesMrB}
\title{Get the Sample of Posterior Trees from a Dated Phylogenetic Analysis with MrBayes (Or a Summary Tree, such as the MCCT)}
\usage{
obtainDatedPosteriorTreesMrB(runFile, nRuns = 2, burnin = 0.5, outputTrees,
  file = NULL)
}
\arguments{
\item{nRuns}{The number of runs in your analysis. This variable is used for figuring out what 
filenames will be searched for: specify less runs than reality and some runs won't be included.
Specify too many, and this function will throw an error when it cannot find files it expects
but do not exist. The default for this argument (two runs) is based on the default number of runs in MrBayes.}

\item{burnin}{The fraction of trees sampled in the posterior discarded  and not returned
by this function directly, nor included in calculation of summary trees. Must be a numeric
value greater than 0 and less than 1.}

\item{outputTrees}{Determines the output trees produced; for format of output, see section
on returned Value below. Must be of length one, and either \code{"all"},
which means all trees from the post-burnin posterior will
returned, a number greater than zero, which will be the number of trees
randomly sampled from across the post-burning posterior and returned,
or \code{"MCCT"} and \code{"MAP"}, which stand for 'maximum clade compatibility tree'
and 'maximum a posteri tree' respectively. The MAP is the single tree from the post-burnin
posterior with the highest marginal likelihood. The MCCT is the single tree from the post-burnin
posterior which contains clades with the highest product of posterior probabilities for its
component clades. Thus, the MAP is the best overall tree, while the MCCT may be the best
tree for summaring topological support.}

\item{file}{Filename (possibly with path) as a character string
leading to a file which will be overwritten with the output trees (or summary tree),
as a NEXUS file. If \code{NULL} (the default), the output will
instead be directly returned by this function.}

\item{runfile}{A filename in the current directory, or a path to a file that is either a .p 
or .t file from a MrBayes analysis. This filename and path will be used for finding additional 
.t and .p files, via the \code{nRuns} settings and assuming that files are in the
same directory \emph{and} these files are named under
typical MrBayes file naming conventions. (In other words, if you've renamed your .p or .t files,
this function probably won't be able to find them.)}
}
\value{
Depending on argument \code{file}, this output is either
returned directly, or instead written out in NEXUS format via
ape's \code{write.NEXUS} function to an external file. The output
will consist either of multiple trees sampled from the post-burnin posterior,
or will consist of a single phylogeny (a summary tree, either the MCCT or the MAP).

Users are warned that the resulting dated trees will not have \code{$root.time} elements
for comparison against an absolute time-scale. Most functions for phylogenetics in R (and
elsewhere) will instead presume that the latest tip will be at time 0 (the modern), which
may be wrong if you are using \code{paleotree} for \emph{paleontological} datasets.
}
\description{
MrBayes is not great for getting samples of dated posterior phylogenies, or for obtaining certain summary trees from
the posterior (specifically the MCCT and MAP, which are specfic trees in the posterior). This is because the tree
samples as returned are scaled relative to rate parameters in a separate file. This function attempts to automate
the handling of multiple files (both .t tree files and .p parameter files), as well as multiple files
associated with separate runs, to obtain samples of posterior trees, or summary trees such as the MCCT or MAP.
}
\details{
This function is most useful for dealing with dating analyses in MrBayes, particularly when tip-dating
a tree with fossil taxa, as the half-compatibility and all-compatibility summary trees offered by the
'sumt' command in MrBayes can hae issues properly portraying summary trees from such datasets.
}
\examples{
\dontrun{

MCCT<-obtainDatedPosteriorTreesMrB(
 	runFile="C:\\\\myTipDatingAnalysis\\\\MrB_run_fossil_05-10-17.nex.run1.t",
 	nRuns=2,burnin=0.5,
		outputTrees="MCCT",file=NULL)

}

}
\seealso{
\code{\link[phangorn]{maxCladeCred}} in package phangorn
}
\author{
David Bapst, with rescaling of raw output
trees via code originally written by Nicholas Crouch.
}
