% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPhylopicTreePBDB.R
\name{plotPhylopicTreePBDB}
\alias{plotPhylopicTreePBDB}
\title{Plot a Phylogeny with Organismal Silhouettes from PhyloPic, Called Via the Paleobiology Database API}
\usage{
plotPhylopicTreePBDB(tree, taxaDataPBDB = taxaTree$taxaData, size = 0.9,
  noiseThreshold = 0.1, extraMargin = 0.2, rescalePNG = TRUE,
  trimPNG = TRUE)
}
\arguments{
\item{tree}{}

\item{taxaDataPBDB}{See \emph{Details}.}

\item{size}{= 0.9}

\item{noiseThreshold}{= 0.1}

\item{extraMargin}{= 0.2}

\item{rescalePNG}{= TRUE}

\item{trimPNG}{= TRUE}
}
\value{
This function returns nothing - it just plots the input tree with
silhouettes from the PhyloPic database.
}
\description{
This function will take a phylogeny, preferably a taxonomy-tree
created from classification information and/or parent-child taxon
information pulled from the Paleobiology Database via function
\code{\link{makePBDBtaxonTree}}, and use the
Paleobiology Database's API to plot silhouettes of each given tip taxon
in replacement of their normal tip labels.
}
\details{
This function preferably will pull the identifiers for which images are to be associated with the tip taxa from \code{taxaDataPBDB$image_no}. By default, \code{taxaDataPBDB} itself is assumed to be an element of \code{tree} named \code{tree$taxaData}, as the PBDB data table used to construct the tree is appended to the output tree when \code{\link{makePBDBtaxonTree}} is used to construct a taxonomy-tree. If the functions listed in \code{\link{getTaxaDataPBDB}} are used to obtain the taxonomic data, this table will include the \code{image_no} variable, which is the image identifier numbers needed to call PNGs from the Paleobiology Database API. If \code{taxaDataPBDB} isn't provided, either by the user directly, or as an element of \code{tree}.
}
\examples{

\donttest{


# now plot with phylopic images
plotPhylopicTreePBDB(tree = tree, 
	taxaDataPBDB = taxaData)

}

}
\references{
Peters, S. E., and M. McClennen. 2015. The Paleobiology Database
application programming interface. \emph{Paleobiology} 42(1):1-7.
}
\seealso{
See \code{\link{getTaxaDataPBDB}}, \code{\link{makePBDBtaxonTree}},
and \code{\link{plotPhylopicTreePBDB}}.
}
\author{
David W. Bapst
}
