% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPhylopicTreePBDB.R
\name{plotPhylopicTreePBDB}
\alias{plotPhylopicTreePBDB}
\title{Plot a Phylogeny with Organismal Silhouettes from PhyloPic, Called Via the Paleobiology Database API}
\usage{
plotPhylopicTreePBDB(tree, taxaDataPBDB = tree$taxaDataPBDB,
  size = 0.9, noiseThreshold = 0.1, extraMargin = 0.2,
  rescalePNG = TRUE, trimPNG = TRUE, makeMonochrome = FALSE,
  cacheDir = NULL, ...)
}
\arguments{
\item{tree}{A phylogeny of class \code{phylo} which will be
plotted, with the terminal tip taxa replaced by silhouettes.
The tree will be plotted with edge lengths.}

\item{taxaDataPBDB}{A \code{data.frame} of taxonomic data from
the Paleobiology Database containing an \code{$image_no} variable,
as returned when \code{show = "img"} is used. See \emph{Details}.}

\item{size}{The default is \code{size = 0.9}.}

\item{noiseThreshold}{A threshold for noise in the PNG from Phylopic
to be treated as meaningless noise (i.e. a color that is effectively
whitespace) and thus can be trimmed as empty margin which can be
trimmed before the silhouette is plotted. The units for this argument
are on a scale from 0 to 1, with 0 being true white space, and values
between 0 and 0.5 representing colors closer to whitespace than true
black. The default is \code{noiseThreshold = 0.1}.}

\item{extraMargin}{The default is \code{extraMargin = 0.2}.}

\item{rescalePNG}{If \code{TRUE} (the default),}

\item{trimPNG}{If \code{TRUE} (the default),}

\item{makeMonochrome}{If \code{TRUE}, PhyloPic silhouettes are
forced to be purely monochrome black-and-white, with no gray
scale. Most of the silhouettes are binary black-and-white already
but some aren't, but those gray-scale values (sometimes?) seem
to exist to indicate very fine features. However, maybe an image
is far too much gray-scale, in which case users can try this
option to force all silhouettes to be monochrome.
The default is \code{FALSE}.}

\item{cacheDir}{If not \code{NULL}, first look here for a cached
version of the images. This makes loading faster.
The default is \code{NULL}.}

\item{...}{Other arguments to pass to plotting code}
}
\value{
This function returns nothing - it just plots the input tree with
silhouettes from the PhyloPic database.
}
\description{
This function will take a phylogeny, preferably a taxonomy-tree
created from classification information and/or parent-child taxon
information pulled from the Paleobiology Database via function
\code{\link{makePBDBtaxonTree}}, and use the
Paleobiology Database's API to plot silhouettes of each given tip taxon
in replacement of their normal tip labels.
}
\details{
This function preferably will pull the identifiers for which images are to be associated with the tip taxa from \code{taxaDataPBDB$image_no}. By default, \code{taxaDataPBDB} itself is assumed to be an element of \code{tree} named \code{tree$taxaData}, as the PBDB data table used to construct the tree is appended to the output tree when \code{\link{makePBDBtaxonTree}} is used to construct a taxonomy-tree. If the functions listed in \code{\link{getTaxaDataPBDB}} are used to obtain the taxonomic data, this table will include the \code{image_no} variable, which is the image identifier numbers needed to call PNGs from the Paleobiology Database API. If \code{taxaDataPBDB} isn't provided, either by the user directly, or as an element of \code{tree}.
}
\examples{

\donttest{


# now plot with phylopic images
plotPhylopicTreePBDB(tree = tree,
	taxaDataPBDB = taxaData)

}

}
\references{
Peters, S. E., and M. McClennen. 2015. The Paleobiology Database
application programming interface. \emph{Paleobiology} 42(1):1-7.
}
\seealso{
See \code{\link{getTaxaDataPBDB}}, \code{\link{makePBDBtaxonTree}},
and \code{\link{plotPhylopicTreePBDB}}.
}
\author{
David W. Bapst
}
