% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rootSplit.R
\name{rootSplit}
\alias{rootSplit}
\title{Split Tip Taxa by Root Divergence}
\usage{
rootSplit(tree)
}
\arguments{
\item{tree}{A phylo object}
}
\value{
Returns a list with each element a character vector containing the
names of terminal taxa descended from each lineage splitting off of the root
node.
}
\description{
Sorts terminal taxa into groups descended from each lineage splitting off of
the root node.
}
\details{
This function can be useful for studying the timing in the order
of appearance of descended from different lineages descended from the first
bifurcation.
}
\examples{

tree<-rtree(100)
rootSplit(tree)

}
\author{
David W. Bapst
}

