% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/testEdgeMat.R
\name{testEdgeMat}
\alias{testEdgeMat}
\title{Test the edge matrix of a phylo object for inconsistencies
This is a small simple function which tests the $edge matrix of 'phylo' objects for
inconsistencies that can cause downstream analytical problems.}
\usage{
testEdgeMat(tree)
}
\arguments{
\item{tree}{A phylogeny object of type phylo}
}
\value{
If all the checks in the function pass correctly, the logical TRUE is returned.
}
\description{
Test the edge matrix of a phylo object for inconsistencies
This is a small simple function which tests the $edge matrix of 'phylo' objects for
inconsistencies that can cause downstream analytical problems.
}
\details{
Useful when doing complex manipulations and reconstitutions of 'phylo' objects (or their
de novo construction), and thus is used by a number of paleotree functions.
}
\examples{
set.seed(444)
tree<-rtree(10)
# should return TRUE
testEdgeMat(tree)
}
\author{
David W. Bapst
}

